pyproject.toml
[tool.poetry]
name = "ksp-proboj-web"
version = "0.1.0"
description = ""
authors = ["Adam Zahradník <adam@zahradnik.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.7"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-widget-tweaks = "^1.5.0"
markdown = "^3.5.1"
pygments = "^2.16.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
